refactor: split core utilities into modules
