+++ lib/index.js
@@ -1,35 +1,42 @@
 #!/usr/bin/env node
+
+// src/index.ts
+import minimist2 from "minimist";
+
+// src/validateEntryPoints.ts
+import {existsSync} from "fs";
+const validateEntryPoints = (options2) => {
+  const missingEntryPoint = options2._.length < 1 && !options2.entry;
+  const invalidEntryPoint = !options2._.every((entryPoint) => existsSync(entryPoint));
+  if (missingEntryPoint)
+    throw new Error("Missing entry point");
+  else if (invalidEntryPoint)
+    throw new Error("Invalid entry point");
+};
+
+// src/watchAndBuild.ts
+import path2 from "path";
 import readline2 from "readline";
-import path2, {dirname} from "path";
+import esbuild2 from "esbuild";
 import {fileURLToPath} from "url";
 import {spawn} from "child_process";
-import minimist2 from "minimist";
-import esbuild2 from "esbuild";
-import {existsSync} from "fs";
-const __dirname = dirname(fileURLToPath(import.meta.url));
-const options = minimist2(process.argv.slice(2));
-const missingEntryPoint = options._.length < 1 && !options.entry;
-const invalidEntryPoint = !options._.every((entryPoint) => existsSync(entryPoint));
-if (missingEntryPoint)
-  throw new Error("Missing entry point");
-else if (invalidEntryPoint)
-  throw new Error("Invalid entry point");
+const __dirname = path2.dirname(fileURLToPath(import.meta.url));
 let child;
 const rl = readline2.createInterface({
   input: process.stdin,
   output: process.stdout
 });
-const watchAndBuild = async () => {
-  const entryPoints = options.entry ? [options.entry] : options._;
-  const outdir = options.outdir || path2.join(__dirname, "build");
+const watchAndBuild = async (options2) => {
+  const entryPoints = options2.entry ? [options2.entry] : options2._;
+  const outdir = options2.outdir || path2.join(__dirname, "build");
   child && child.kill();
   const service = await esbuild2.startService();
-  options.clear && console.clear();
+  options2.clear && console.clear();
   await service.build({entryPoints, outdir});
   return new Promise((resolve) => {
-    if (options.run) {
+    if (options2.run) {
       rl.pause();
-      const commandToRun = typeof options.run === "string" ? options.run : `node ${outdir}`;
+      const commandToRun = typeof options2.run === "string" ? options2.run : `node ${outdir}`;
       const commandName = commandToRun.split(" ").shift();
       const commandParameters = commandToRun.split(" ").slice(1);
       child = spawn(commandName, commandParameters, {stdio: "inherit"});
@@ -42,4 +49,8 @@ const watchAndBuild = async () => {
     }
   });
 };
-options.watch ? import("chokidar").then(({default: chokidar}) => chokidar.watch(options.watch).on("ready", async () => watchAndBuild()).on("change", async () => watchAndBuild())) : watchAndBuild();
+
+// src/index.ts
+const options = minimist2(process.argv.slice(2));
+validateEntryPoints(options);
+options.watch ? import("chokidar").then(({default: chokidar}) => chokidar.watch(options.watch).on("ready", async () => watchAndBuild(options)).on("change", async () => watchAndBuild(options))) : watchAndBuild(options);
