+++ src/types.ts
@@ -1,0 +1,9 @@
+import type { ParsedArgs } from 'minimist'
+
+export interface CLIFlags extends ParsedArgs {
+  watch: string | readonly string[]
+  clear?: boolean
+  entry: string
+  outdir?: string
+  run?: boolean
+}
